Split plan expression shapes
- Move result-family expression structs into dedicated plan expression modules.
- Keep the root expression module focused on Expr, CallArg, and cross-family conversions.
- Keep public expression exports unchanged.
- Move kind accessor coverage next to each result-family module.
