+++ src/plan/expression.rs
@@ -1,11 +1,20 @@
-use super::id::{
-    BoolFunctionId, BoolLocalId, IntFunctionId, IntLocalId, LocalId, NilFunctionId, NilLocalId,
-    StringFunctionId, StringLocalId,
-};
-use super::step::Step;
-use super::value::{FunctionType, FunctionValue, Value, ValueType};
-use ecow::EcoString;
+mod bool;
+mod function;
+mod int;
+mod nil;
+mod string;
+
+use super::id::{BoolLocalId, IntLocalId, LocalId, NilLocalId, StringLocalId};
+use super::value::{Value, ValueType};
 use num_bigint::BigInt;
+
+pub use self::{
+    bool::BoolExpr, function::FunctionExpr, int::IntExpr, nil::NilExpr, string::StringExpr,
+};
+pub(crate) use self::{
+    bool::BoolExprKind, function::FunctionExprKind, int::IntExprKind, nil::NilExprKind,
+    string::StringExprKind,
+};
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct Expr {
@@ -43,217 +52,6 @@ pub(crate) enum CallArgKind {
     Nil {
         local: NilLocalId,
         value: NilExpr,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct IntExpr {
-    kind: IntExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum IntExprKind {
-    Value(BigInt),
-    LocalGet {
-        local: IntLocalId,
-        name: EcoString,
-    },
-    Call {
-        function: IntFunctionId,
-        args: Vec<CallArg>,
-    },
-    Add {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    Sub {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    Mult {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    Div {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    Remainder {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    Negate(Box<IntExpr>),
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<IntExpr>,
-        false_: Box<IntExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, IntExpr)>,
-        fallback: Box<IntExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<IntExpr>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct StringExpr {
-    kind: StringExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum StringExprKind {
-    Value(EcoString),
-    LocalGet {
-        local: StringLocalId,
-        name: EcoString,
-    },
-    Call {
-        function: StringFunctionId,
-        args: Vec<CallArg>,
-    },
-    Concatenate {
-        left: Box<StringExpr>,
-        right: Box<StringExpr>,
-    },
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<StringExpr>,
-        false_: Box<StringExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, StringExpr)>,
-        fallback: Box<StringExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<StringExpr>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct BoolExpr {
-    kind: BoolExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum BoolExprKind {
-    Value(bool),
-    LocalGet {
-        local: BoolLocalId,
-        name: EcoString,
-    },
-    Call {
-        function: BoolFunctionId,
-        args: Vec<CallArg>,
-    },
-    Not(Box<BoolExpr>),
-    LtInt {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    LtEqInt {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    GtInt {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    GtEqInt {
-        left: Box<IntExpr>,
-        right: Box<IntExpr>,
-    },
-    Equal {
-        left: Box<Expr>,
-        right: Box<Expr>,
-    },
-    NotEqual {
-        left: Box<Expr>,
-        right: Box<Expr>,
-    },
-    And {
-        left: Box<BoolExpr>,
-        right: Box<BoolExpr>,
-    },
-    Or {
-        left: Box<BoolExpr>,
-        right: Box<BoolExpr>,
-    },
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<BoolExpr>,
-        false_: Box<BoolExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, BoolExpr)>,
-        fallback: Box<BoolExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<BoolExpr>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct NilExpr {
-    kind: NilExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum NilExprKind {
-    Value,
-    LocalGet {
-        local: NilLocalId,
-        name: EcoString,
-    },
-    Call {
-        function: NilFunctionId,
-        args: Vec<CallArg>,
-    },
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<NilExpr>,
-        false_: Box<NilExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, NilExpr)>,
-        fallback: Box<NilExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<NilExpr>,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct FunctionExpr {
-    type_: FunctionType,
-    kind: FunctionExprKind,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) enum FunctionExprKind {
-    Value(FunctionValue),
-    BoolCase {
-        subject: Box<BoolExpr>,
-        true_: Box<FunctionExpr>,
-        false_: Box<FunctionExpr>,
-    },
-    IntCase {
-        subject: Box<IntExpr>,
-        clauses: Vec<(BigInt, FunctionExpr)>,
-        fallback: Box<FunctionExpr>,
-    },
-    Block {
-        steps: Vec<Step>,
-        return_: Box<FunctionExpr>,
     },
 }
 
@@ -498,425 +296,6 @@ impl CallArg {
     }
 }
 
-impl IntExpr {
-    pub(crate) fn value(value: BigInt) -> Self {
-        Self {
-            kind: IntExprKind::Value(value),
-        }
-    }
-
-    pub(crate) fn local_get(local: IntLocalId, name: EcoString) -> Self {
-        Self {
-            kind: IntExprKind::LocalGet { local, name },
-        }
-    }
-
-    pub(crate) fn call(function: IntFunctionId, args: Vec<CallArg>) -> Self {
-        Self {
-            kind: IntExprKind::Call { function, args },
-        }
-    }
-
-    pub(crate) fn add(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: IntExprKind::Add {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn sub(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: IntExprKind::Sub {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn mult(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: IntExprKind::Mult {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn div(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: IntExprKind::Div {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn remainder(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: IntExprKind::Remainder {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn negate(value: IntExpr) -> Self {
-        Self {
-            kind: IntExprKind::Negate(Box::new(value)),
-        }
-    }
-
-    pub(crate) fn bool_case(subject: BoolExpr, true_: IntExpr, false_: IntExpr) -> Self {
-        Self {
-            kind: IntExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, IntExpr)>,
-        fallback: IntExpr,
-    ) -> Self {
-        Self {
-            kind: IntExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: IntExpr) -> Self {
-        Self {
-            kind: IntExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub(crate) fn kind(&self) -> &IntExprKind {
-        &self.kind
-    }
-}
-
-impl StringExpr {
-    pub(crate) fn value(value: EcoString) -> Self {
-        Self {
-            kind: StringExprKind::Value(value),
-        }
-    }
-
-    pub(crate) fn local_get(local: StringLocalId, name: EcoString) -> Self {
-        Self {
-            kind: StringExprKind::LocalGet { local, name },
-        }
-    }
-
-    pub(crate) fn call(function: StringFunctionId, args: Vec<CallArg>) -> Self {
-        Self {
-            kind: StringExprKind::Call { function, args },
-        }
-    }
-
-    pub(crate) fn concatenate(left: StringExpr, right: StringExpr) -> Self {
-        Self {
-            kind: StringExprKind::Concatenate {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn bool_case(subject: BoolExpr, true_: StringExpr, false_: StringExpr) -> Self {
-        Self {
-            kind: StringExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, StringExpr)>,
-        fallback: StringExpr,
-    ) -> Self {
-        Self {
-            kind: StringExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: StringExpr) -> Self {
-        Self {
-            kind: StringExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub(crate) fn kind(&self) -> &StringExprKind {
-        &self.kind
-    }
-}
-
-impl BoolExpr {
-    pub(crate) fn value(value: bool) -> Self {
-        Self {
-            kind: BoolExprKind::Value(value),
-        }
-    }
-
-    pub(crate) fn local_get(local: BoolLocalId, name: EcoString) -> Self {
-        Self {
-            kind: BoolExprKind::LocalGet { local, name },
-        }
-    }
-
-    pub(crate) fn call(function: BoolFunctionId, args: Vec<CallArg>) -> Self {
-        Self {
-            kind: BoolExprKind::Call { function, args },
-        }
-    }
-
-    pub(crate) fn not(value: BoolExpr) -> Self {
-        Self {
-            kind: BoolExprKind::Not(Box::new(value)),
-        }
-    }
-
-    pub(crate) fn lt_int(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: BoolExprKind::LtInt {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn lte_int(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: BoolExprKind::LtEqInt {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn gt_int(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: BoolExprKind::GtInt {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn gte_int(left: IntExpr, right: IntExpr) -> Self {
-        Self {
-            kind: BoolExprKind::GtEqInt {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn equal(left: Expr, right: Expr) -> Self {
-        Self {
-            kind: BoolExprKind::Equal {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn not_equal(left: Expr, right: Expr) -> Self {
-        Self {
-            kind: BoolExprKind::NotEqual {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn and(left: BoolExpr, right: BoolExpr) -> Self {
-        Self {
-            kind: BoolExprKind::And {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn or(left: BoolExpr, right: BoolExpr) -> Self {
-        Self {
-            kind: BoolExprKind::Or {
-                left: Box::new(left),
-                right: Box::new(right),
-            },
-        }
-    }
-
-    pub(crate) fn bool_case(subject: BoolExpr, true_: BoolExpr, false_: BoolExpr) -> Self {
-        Self {
-            kind: BoolExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, BoolExpr)>,
-        fallback: BoolExpr,
-    ) -> Self {
-        Self {
-            kind: BoolExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: BoolExpr) -> Self {
-        Self {
-            kind: BoolExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub(crate) fn kind(&self) -> &BoolExprKind {
-        &self.kind
-    }
-}
-
-impl NilExpr {
-    pub(crate) fn value() -> Self {
-        Self {
-            kind: NilExprKind::Value,
-        }
-    }
-
-    pub(crate) fn local_get(local: NilLocalId, name: EcoString) -> Self {
-        Self {
-            kind: NilExprKind::LocalGet { local, name },
-        }
-    }
-
-    pub(crate) fn call(function: NilFunctionId, args: Vec<CallArg>) -> Self {
-        Self {
-            kind: NilExprKind::Call { function, args },
-        }
-    }
-
-    pub(crate) fn bool_case(subject: BoolExpr, true_: NilExpr, false_: NilExpr) -> Self {
-        Self {
-            kind: NilExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, NilExpr)>,
-        fallback: NilExpr,
-    ) -> Self {
-        Self {
-            kind: NilExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: NilExpr) -> Self {
-        Self {
-            kind: NilExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub(crate) fn kind(&self) -> &NilExprKind {
-        &self.kind
-    }
-}
-
-impl FunctionExpr {
-    pub(crate) fn value(value: FunctionValue) -> Self {
-        Self {
-            type_: value.type_(),
-            kind: FunctionExprKind::Value(value),
-        }
-    }
-
-    pub(crate) fn bool_case(subject: BoolExpr, true_: FunctionExpr, false_: FunctionExpr) -> Self {
-        Self {
-            type_: true_.type_.clone(),
-            kind: FunctionExprKind::BoolCase {
-                subject: Box::new(subject),
-                true_: Box::new(true_),
-                false_: Box::new(false_),
-            },
-        }
-    }
-
-    pub(crate) fn int_case(
-        subject: IntExpr,
-        clauses: Vec<(BigInt, FunctionExpr)>,
-        fallback: FunctionExpr,
-    ) -> Self {
-        Self {
-            type_: fallback.type_.clone(),
-            kind: FunctionExprKind::IntCase {
-                subject: Box::new(subject),
-                clauses,
-                fallback: Box::new(fallback),
-            },
-        }
-    }
-
-    pub(crate) fn block(steps: Vec<Step>, return_: FunctionExpr) -> Self {
-        Self {
-            type_: return_.type_.clone(),
-            kind: FunctionExprKind::Block {
-                steps,
-                return_: Box::new(return_),
-            },
-        }
-    }
-
-    pub fn type_(&self) -> &FunctionType {
-        &self.type_
-    }
-
-    pub(crate) fn kind(&self) -> &FunctionExprKind {
-        &self.kind
-    }
-}
-
 impl From<Value> for Expr {
     fn from(value: Value) -> Self {
         match value {
@@ -931,13 +310,10 @@ impl From<Value> for Expr {
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        BoolExpr, BoolExprKind, CallArgKind, Expr, ExprKind, FunctionExpr, FunctionExprKind,
-        IntExpr, IntExprKind, NilExpr, NilExprKind, StringExpr, StringExprKind,
-    };
+    use super::{BoolExpr, CallArgKind, Expr, FunctionExpr, IntExpr, NilExpr, StringExpr};
     use crate::plan::{
         BoolLocalId, FunctionType, FunctionValue, IntFunctionId, IntLocalId, LocalId, NilLocalId,
-        RuntimeFunctionId, Step, StringFunctionId, StringLocalId, Value, ValueType,
+        RuntimeFunctionId, StringFunctionId, StringLocalId, Value, ValueType,
     };
     use num_bigint::BigInt;
 
@@ -1185,164 +561,6 @@ mod tests {
             Expr::from(Value::Function(function_value())).value_type(),
             ValueType::Function(Box::new(function_type())),
         );
-    }
-
-    #[test]
-    fn typed_expression_kind_accessors() {
-        assert!(matches!(
-            IntExpr::value(BigInt::from(1)).kind(),
-            IntExprKind::Value(_)
-        ));
-        assert!(matches!(
-            StringExpr::value("geam".into()).kind(),
-            StringExprKind::Value(_)
-        ));
-        assert!(matches!(
-            BoolExpr::value(true).kind(),
-            BoolExprKind::Value(true)
-        ));
-        assert!(matches!(NilExpr::value().kind(), NilExprKind::Value));
-        assert!(matches!(
-            IntExpr::bool_case(
-                BoolExpr::value(true),
-                IntExpr::value(1.into()),
-                IntExpr::value(0.into())
-            )
-            .kind(),
-            IntExprKind::BoolCase { .. }
-        ));
-        assert!(matches!(
-            IntExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), IntExpr::value(10.into()))],
-                IntExpr::value(0.into())
-            )
-            .kind(),
-            IntExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            IntExpr::block(
-                vec![Step::evaluate(Expr::int(IntExpr::value(1.into())))],
-                IntExpr::value(2.into()),
-            )
-            .kind(),
-            IntExprKind::Block { .. }
-        ));
-        assert!(matches!(
-            StringExpr::bool_case(
-                BoolExpr::value(true),
-                StringExpr::value("yes".into()),
-                StringExpr::value("no".into())
-            )
-            .kind(),
-            StringExprKind::BoolCase { .. }
-        ));
-        assert!(matches!(
-            StringExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), StringExpr::value("one".into()))],
-                StringExpr::value("other".into())
-            )
-            .kind(),
-            StringExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            StringExpr::block(
-                vec![Step::evaluate(Expr::string(StringExpr::value("a".into())))],
-                StringExpr::value("b".into()),
-            )
-            .kind(),
-            StringExprKind::Block { .. }
-        ));
-        assert!(matches!(
-            BoolExpr::bool_case(
-                BoolExpr::value(true),
-                BoolExpr::value(true),
-                BoolExpr::value(false)
-            )
-            .kind(),
-            BoolExprKind::BoolCase { .. }
-        ));
-        assert!(matches!(
-            BoolExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), BoolExpr::value(true))],
-                BoolExpr::value(false)
-            )
-            .kind(),
-            BoolExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            BoolExpr::and(BoolExpr::value(true), BoolExpr::value(false)).kind(),
-            BoolExprKind::And { .. }
-        ));
-        assert!(matches!(
-            BoolExpr::or(BoolExpr::value(true), BoolExpr::value(false)).kind(),
-            BoolExprKind::Or { .. }
-        ));
-        assert!(matches!(
-            BoolExpr::block(
-                vec![Step::evaluate(Expr::bool(BoolExpr::value(false)))],
-                BoolExpr::value(true),
-            )
-            .kind(),
-            BoolExprKind::Block { .. }
-        ));
-        assert!(matches!(
-            NilExpr::bool_case(BoolExpr::value(true), NilExpr::value(), NilExpr::value()).kind(),
-            NilExprKind::BoolCase { .. }
-        ));
-        assert!(matches!(
-            NilExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), NilExpr::value())],
-                NilExpr::value()
-            )
-            .kind(),
-            NilExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            NilExpr::block(
-                vec![Step::evaluate(Expr::nil(NilExpr::value()))],
-                NilExpr::value(),
-            )
-            .kind(),
-            NilExprKind::Block { .. }
-        ));
-        assert!(matches!(
-            Expr::from(Value::Nil).kind(),
-            ExprKind::Nil(NilExpr { .. })
-        ));
-        assert!(matches!(
-            FunctionExpr::value(function_value()).kind(),
-            FunctionExprKind::Value(_)
-        ));
-        assert!(matches!(
-            FunctionExpr::int_case(
-                IntExpr::value(1.into()),
-                vec![(1.into(), FunctionExpr::value(function_value()))],
-                FunctionExpr::value(function_value()),
-            )
-            .kind(),
-            FunctionExprKind::IntCase { .. }
-        ));
-        assert!(matches!(
-            FunctionExpr::block(
-                vec![Step::evaluate(Expr::int(IntExpr::value(1.into())))],
-                FunctionExpr::value(function_value()),
-            )
-            .kind(),
-            FunctionExprKind::Block { .. }
-        ));
-        assert!(matches!(
-            FunctionExpr::bool_case(
-                BoolExpr::value(true),
-                FunctionExpr::value(function_value()),
-                FunctionExpr::value(function_value()),
-            )
-            .kind(),
-            FunctionExprKind::BoolCase { .. }
-        ));
     }
 
     #[test]
